#14 [Fix] 북마크 삭제 오류 수정

+++ src/main/java/com/mpnp/baechelin/review/service/ReviewService.java
@@ -53,10 +53,10 @@ public class ReviewService {
      */
     public void review(ReviewRequestDto reviewRequestDto, String socialId) throws IOException {
 
-        long   storeId = reviewRequestDto.getStoreId();
-        Store  store   = storeRepository.findById(storeId).orElseThrow(() -> new IllegalArgumentException("해당하는 업장이 존재하지 않습니다."));
-        User   user    = userRepository.findBySocialId(socialId);
-        Review review  = new Review(reviewRequestDto, store, user);
+        long storeId = reviewRequestDto.getStoreId();
+        Store store = storeRepository.findById(storeId).orElseThrow(() -> new IllegalArgumentException("해당하는 업장이 존재하지 않습니다."));
+        User user = userRepository.findBySocialId(socialId);
+        Review review = new Review(reviewRequestDto, store, user);
 
 
         // todo 태크 매핑
@@ -72,7 +72,7 @@ public class ReviewService {
 
 
         // todo 이미지가 널값이 아니라면 업로드 실행
-        if (newReviewImage != null) {
+        if (newReviewImage != null && !newReviewImage.isEmpty()) {
             for (MultipartFile reviewImageFile : newReviewImage) {
                 String fileDir = awsS3Manager.uploadFile(reviewImageFile);
                 log.info("upload --> " + fileDir);
@@ -215,7 +215,7 @@ public class ReviewService {
         Optional<Review> review = reviewRepository.findById(reviewId);      // 리뷰 매핑
 
         if (user == null) {
-           throw new IllegalArgumentException("해당하는 소셜아이디를 찾을 수 없습니다.");
+            throw new IllegalArgumentException("해당하는 소셜아이디를 찾을 수 없습니다.");
         }      // 유저 유무 확인 예외처리
 
         review.orElseThrow(() -> new IllegalArgumentException("해당하는 리뷰가 이미 삭제 되었습니다.")); // 리뷰 유무 확인 예외처리
@@ -227,6 +227,7 @@ public class ReviewService {
         reviewRepository.deleteById(review.get().getId()); // 1
         if (!review.get().getReviewImageList().isEmpty()) { // 2
             for (ReviewImage reviewImage : imageList) {
+                if (reviewImage.getReviewImageUrl() == null || reviewImage.getReviewImageUrl().equals("")) continue;
                 System.out.println("delete -> " + reviewImage.getReviewImageUrl().substring(reviewImage.getReviewImageUrl().indexOf("com/") + 4));
                 awsS3Manager.deleteFile(reviewImage.getReviewImageUrl().substring(reviewImage.getReviewImageUrl().indexOf("com/") + 4));
             }
